refactor: use direct API type imports
